requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
